% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_ind.R
\name{calcul_ind}
\alias{calcul_ind}
\title{Tableau des calculs usuels}
\usage{
calcul_ind(X, Y, N)
}
\arguments{
\item{N}{the number of individual}

\item{x}{the X-axis coordinate of an individual}

\item{y}{the Y-axis coordinate of an individual}
}
\value{
a table of usual indicators
}
\description{
Tableau des calculs usuels
}
\examples{
calcul_ind(data1$X,data1$Y,nrow(data1))
}
