% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRect.R
\name{covRect}
\alias{covRect}
\title{Représentation de la covariance}
\usage{
covRect(data, rank = FALSE, timeseries = FALSE)
}
\arguments{
\item{data}{a dataset}
}
\value{
A mist of GGPLOT graph explainig
}
\description{
Représentation de la covariance
}
\examples{
covRect(data1)[[1]]
}
