% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covRect_resample.R
\name{covRect_resample}
\alias{covRect_resample}
\title{Représentation de la covariance après réchantillonage}
\usage{
covRect_resample(datas, size)
}
\arguments{
\item{size}{size of the dataset after resampling}

\item{data}{a dataset}
}
\value{
A mist of GGPLOT graph following resampling
}
\description{
Représentation de la covariance après réchantillonage
}
\examples{
covRect_resample(data1)[[1]]
}
