% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{data1}
\alias{data1}
\title{A sample with randomly generate with strong variance of Y}
\format{
A data frame with 500 rows:
\describe{
\item{X}{sort(rnorm(N, 40, 100)) }
\item{Y}{1- 2*X + rnorm(N, 0, 200) # forte variance}
...
}
}
\usage{
data1
}
\description{
A dataset containing the coordonates of 500 point
}
\keyword{datasets}
