% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphCor.R
\name{graph_cor}
\alias{graph_cor}
\title{Représentation de corrélation}
\usage{
graph_cor(X, Y, N)
}
\arguments{
\item{N}{the number of individual}

\item{x}{the X-axis coordinate of an individual}

\item{y}{the Y-axis coordinate of an individual}
}
\value{
A GGPLOT graph
}
\description{
Représentation de corrélation
}
\examples{
graph_cor(data1$X,data1$Y,nrow(data1))
}
