% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlsDiag.R
\name{OlsDiag}
\alias{OlsDiag}
\title{Représentation de OLS 2 variables}
\usage{
OlsDiag(X, Y, N, with_OLS = TRUE)
}
\arguments{
\item{X}{the X-axis coordinate of an individual}

\item{Y}{the Y-axis coordinate of an individual}

\item{N}{the number of individual}
}
\value{
A GGPLOT graph explainig OLS
}
\description{
Représentation de OLS 2 variables
}
\examples{
OlsDiag(data1$X,data1$Y1,nrow(data1))
}
