% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testlin.R
\name{testlin}
\alias{testlin}
\title{Graphique pour test de linéarité}
\usage{
testlin(data)
}
\arguments{
\item{data}{a dataset}
}
\value{
a set GGPLOT graph
}
\description{
Graphique pour test de linéarité
}
\examples{
testlin(data1)
}
